Move agent server modules into agent directory
