+++ apps/web/src/app/.well-known/agent.json/route.ts
@@ -1,7 +1,7 @@
 /**
  * Optional mirror of the agent discovery document.
  */
-import { buildAgentDiscoveryEnvelope } from "@/server/agentDiscovery";
+import { buildAgentDiscoveryEnvelope } from "@/server/agent/discovery";
 
 export const GET = async (request: Request): Promise<Response> =>
   Response.json(buildAgentDiscoveryEnvelope(request));
