+++ apps/web/src/app/api/agent-me-route.test.ts
@@ -1,7 +1,7 @@
 import assert from "node:assert/strict";
 import test from "node:test";
 import { getAgentMeRouteWithDeps } from "./agent/me/route";
-import type { AgentAuthenticatedRequest } from "../../server/agentApiKeyAuth";
+import type { AgentAuthenticatedRequest } from "@/server/agent/apiKeyAuth";
 
 const createAuthenticatedRequest = (): AgentAuthenticatedRequest => ({
   transport: "api_key",
