+++ apps/web/src/app/api/agent/route.ts
@@ -4,7 +4,7 @@
  * This endpoint is intentionally public so an agent can understand the service,
  * bootstrap auth on auth.*, and learn the next steps without prior knowledge.
  */
-import { buildAgentDiscoveryEnvelope } from "@/server/agentDiscovery";
+import { buildAgentDiscoveryEnvelope } from "@/server/agent/discovery";
 
 export const GET = async (request: Request): Promise<Response> =>
   Response.json(buildAgentDiscoveryEnvelope(request));
